only display any given window once.
This change goes along with the sticky window work.  It makes xmonad
display each window once and only once, with preference given to the
focussed screen.  It has no effect when there are no duplicate windows,
except to make things less efficient.  We could do better using Data.Set
(or Data.Map) to store the set of windows that are visible.