Use current workspace to set dimensions of unmapped floating windows
This fixes a bug when using multiple screens with different dimensions, causing floating windows to always increase in size or always decrease in size every time they're opened.

In such a situation, a floating window is first mapped at (0,0) (on the leftmost screen), using the dimensions of that screen, before being inserted in the current workspace, which may be on another screen, causing the window to be resized if that screen has larger or smaller dimensions. If the window then saves its size in pixels when closed, the next time it opens, it will go through this process again, causing it to become even larger or smaller.

This commit fixes this by using the current workspace to calculate dimensions for unmapped floating windows, instead of the workspace at screen position (0,0).