Shift the logHook into handleRefresh before render
Using `handleRefresh` as we are, a longstanding issue with the `logHook`
is fixed, or rather, transformed: running `windows` within it will not
cause an infinite loop; any requests it issues are silently discarded.
That, however, is not ideal either: unhandled changes may be introduced
to the windowset.

We could run the `logHook` via `handleRefresh`, but that would put us
back where we started with infinite loops. Alternatively, we could treat
a request in the `logHook` as an error and discard its changes à la
`catchX`.

However, the way that gives the best bang for buck is to move the
`logHook` from `render` into `handleRefresh`, after the user effects
have run but before the refresh itself. Like this, any changes it makes
are seamlessly handled in the refresh it accompanies. This is simple,
efficient, and best of all: it transforms a mere `logHook` into a fully
fledged "refreshHook".

While the `logHook` is seeing essentially the same state as far as the
internal model goes, there's one key observable difference: it runs
before the effects in `render`, and thus it does not see any changes
that might result from running the layout. This could break some
existing usage.