Initialize XMONAD_* env variables separatly
The user shouldn’t have to input all three variables if they just want
to set where e.g. the config is.

This tries to achive that, but of course we need to keep the
heuristics for the `xmonad` config dir, so set the data dir here if
any of the config/cache files exist and it was not overwritten.